requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.2
djangorestframework-simplejwt==5.0.0
pytz==2021.1
sqlparse==0.4.1
waitress==1.4.4
whitenoise==5.2.0
django-cors-headers==3.9.0
Pillow==8.3.1
PyMySQL==1.0.2


